+++ src/app/_components/modal/columnManagerModal/ColumnManageModal.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Input from '@/app/_components/Input';
-import ModalContainer from '@/app/_components/modal/ModalContainer';
+import ModalContainer from '@/app/_components/modal/modalContainer/ModalContainer';
 import { ModalPropsType } from '@/app/_types/modalProps';
-import CheckCancleButton from './CheckCancleButton';
+import CheckCancleButton from '../checkCancleButton/CheckCancleButton';
 
 const ColumnManageModal = ({
   openModalType,
