+++ src/app/_components/modal/deleteColumnCheckModal/DeleteColumnCheckModal.module.css
