+++ src/app/_components/modal/modalPortal/ModalPortal.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { createPortal } from 'react-dom';
 import { ModalPortalPropsType } from '@/app/_types/modalProps';
 import styles from './ModalPortal.module.css';
-import ModalContents from './ModalContents';
+import ModalContents from '../ModalContents';
 
 const ModalPortal = ({
   openModalType,
