+++ src/app/_components/modal/newColumnModal/NewColumnModal.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import ModalContainer from '@/app/_components/modal/ModalContainer';
+import ModalContainer from '@/app/_components/modal/modalContainer/ModalContainer';
 import { ModalPropsType } from '@/app/_types/modalProps';
 import Input from '@/app/_components/Input';
-import CheckCancleButton from './CheckCancleButton';
+import CheckCancleButton from '../checkCancleButton/CheckCancleButton';
 
 const NewColumnModal = ({ setOpenModalType }: ModalPropsType) => {
   const produceButtonHandler = () => {
